+++ packages/map-template/src/atoms/shuttleBusOnState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const shuttleBusOnState = atom({
+    key: 'shuttleBusOn',
+    default: false
+});
+
+export default shuttleBusOnState;