+++ packages/map-template/src/components/MapsIndoorsMap/MapsIndoorsMap.jsx
@@ -37,7 +37,8 @@ MapsIndoorsMap.propTypes = {
     center: PropTypes.string,
     useAppTitle: PropTypes.bool,
     showMapMarkers: PropTypes.bool,
-    mapboxMapStyle: PropTypes.string
+    mapboxMapStyle: PropTypes.string,
+    devicePosition: PropTypes.object
 };
 
 /**
@@ -166,6 +167,7 @@ function MapsIndoorsMap(props) {
             miTransitionLevel: props.supportsUrlParameters && miTransitionLevelQueryParameter ? miTransitionLevelQueryParameter : props.miTransitionLevel,
             category: props.supportsUrlParameters && categoryQueryParameter ? categoryQueryParameter : props.category,
             center: props.supportsUrlParameters && centerQueryParameter ? centerQueryParameter : props.center,
+            devicePosition: props.devicePosition,
             // Handle boolean values
             useKeyboard: getBooleanValue(props.supportsUrlParameters, defaultProps.useKeyboard, props.useKeyboard, useKeyboardQueryParameter),
             useMapProviderModule: getBooleanValue(props.supportsUrlParameters, defaultProps.useMapProviderModule, props.useMapProviderModule, useMapProviderModuleQueryParameter),